Error out on invalid JSON
This was something small that I lost some time on today, I had a JSON
file with:

```
{
  "bad": "if(bb > 9, 2, 2)",
  "bb": "10",
}
```

Which looked to be valid JSON (it is not, see the trailing comma) and
hesabu would give me the following error:

```
  "errors": [
    {
      "source": "general",
      "expression": "general",
      "message": "cycle between equations"
    }
  ]
```

This was caused because the length of the equations was 0 and the
reverse sort of an empty array is an empty array, so we thought we had
a cycle.

I know error earlier to makes this more clear.