Remove regexp from cleaning function
Turns out, regex in golang is slow. When we were using the
regexp.ReplaceAll we noticed a regression in our performance. Running
the large_equations test set was about a second slower than before.

With strings.Replace we didn't have that drawback, a new function was
needed for the single equals because that needs to additional checks
to see if it is a single = that needs replacing to a ==.