Fix inequality
Or at least support inequality in the formulas. I was removing the
cleaner spec from the Ruby hesabu gem, when I spotted that formulas
could also use `!=`, the cleaner version here did not expect this and
would make it `!==`. This fixes that.

It also brings over the other specs from the ruby gem (the cleaner
part) that were not addressed by the go tests.