Add profiling capability
You can run it with

❯ bin/hesabucli -cpuprofile cpu.prof -memprofile mem.prof bla.json

(for a version with 60000 equations see
https://pile.pjaspers.com/bla-e6zB9y.json )

Now you should have  bot a cpu.prof and a mem.prof in your directory.

If you now run:

go tool pprof cpu.prof

You'll be dropped into an interactive profile viewer, I don't know
much about it, but if I type `web` it will generate a svg.

Currently it looks like this:

https://pile.pjaspers.com/pprof001-GS0e21.svg

Even cooler is the interactive version that opens in your browser:

❯ pprof -http=localhost:1234 bin/hesabucli cpu.prof