Add an `array` function
This makes it possible to use `array` in functions, which is
particularly handy when combined with `eval_array`:

      sum(eval_array('a', array(1,2), 'b', array(2,3), 'a + b'))

This makes it clear that it uses two arrays, the go part didn't really
need this conversion (hence the sparse implementation of that
function) but the ruby/dentaku part did need it. To keep parity
between the two, the array function got added here as well.

This also refactors our tests a bit, to check for more things and be
more verbose when things are not correct