Handle 0/0
I ran into a situation where orbf2 could not run a simulation, it just
gave an error while processing the JSON coming out of Hesabu. When I
investigated it seemed to return just an empty string.

Some further digging pointed me to the JSON-generation failing because
it encountered an error (that's why it will now print that error, so
we have less digging)

We handled Infinity (divide by zero), but 0/0 is a special case and
returns NaN, when we try to generate JSON from NaN, things go not so
well. So I've made that an error that:

- Points to the equation
- Explains what is going on
- Is tested