Add missing upcased variants (strlen and eval_array)
Turns out we were also missing an upcase variant of `strlen`, so I've
added that as well.

I've also moved most of the functions to the static one and sorted
them alphabetically (regardless of case) so we can spot misses more
easily.

The Functions() will now use the static one and just append the two
that eval_array can't use.